# website/openapi/engine.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperatorRecoveryRequest": {
        "description": "Body for operator retry-step / retry-compensation.",
        "properties": {
          "allow_destructive": {
            "default": false,
            "description": "Required with force on commit steps (duplicate side-effect risk).",
            "title": "Allow Destructive",
            "type": "boolean"
          },
          "force": {
            "default": false,
            "description": "Release a non-stale worker claim (reason steps; commit requires allow_destructive).",
            "title": "Force",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional operator note for audit hooks.",
            "title": "Reason"
          },
          "recovery_token": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional client-supplied token (8\u2013128 chars). When provided, duplicate recovery requests with the same token and parameters return the original 202 response without re-enqueueing work.",
            "title": "Recovery Token"
          }
        },
        "title": "OperatorRecoveryRequest",
        "type": "object"
      },
...
      "RecoveryResponse": {
        "description": "Response for operator recovery enqueue (202 Accepted).",
        "properties": {
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "recovery_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recovery Token"
          },
          "status": {
            "enum": [
              "ok",
              "scheduled",
              "claim_active",
              "requeued"
            ],
            "title": "Status",
            "type": "string"
          },
          "worker_command_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Command Key"
          }
        },
        "required": [
          "status"
        ],
        "title": "RecoveryResponse",
        "type": "object"
      },
...
      "SagaInstanceItem": {
        "description": "One row in GET /v1/sagas (instance list).",
        "properties": {
          "definition_id": {
            "title": "Definition Id",
            "type": "string"
          },
          "definition_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition Name"
          },
          "definition_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition Version"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "parent_trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Trace Id"
          },
          "start_idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Idempotency Key"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "namespace",
          "definition_id",
          "status",
          "started_at"
        ],
        "title": "SagaInstanceItem",
        "type": "object"
      },
      "SagaInstanceListResponse": {
        "description": "Response for GET /v1/sagas.",
        "properties": {
          "has_more": {
            "description": "True when len(items) == limit (more pages may exist).",
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SagaInstanceItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Total matching rows when include_total=true was requested.",
            "title": "Total"
          }
        },
        "required": [
          "items",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "SagaInstanceListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Control-plane HTTP API for deploying manifests, starting sagas, HITL review, and operator recovery. OSS surface under /v1.",
    "title": "API Reference",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/sagas": {
      "get": {
        "description": "List saga instances, newest started_at first.",
        "operationId": "get_sagas_v1_sagas_get",
        "parameters": [
          {
            "description": "Filter by namespace; omit for all.",
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by namespace; omit for all.",
              "title": "Namespace"
            }
          },
          {
            "description": "Filter to a single saga instance trace_id.",
            "in": "query",
            "name": "trace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to a single saga instance trace_id.",
              "title": "Trace Id"
            }
          },
          {
            "description": "Filter to child sagas spawned by this parent trace_id.",
            "in": "query",
            "name": "parent_trace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to child sagas spawned by this parent trace_id.",
              "title": "Parent Trace Id"
            }
          },
          {
            "description": "Repeat for multiple values, e.g. status=RUNNING&status=PENDING.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeat for multiple values, e.g. status=RUNNING&status=PENDING.",
              "title": "Status"
            }
          },
          {
            "description": "When true, filter to non-terminal in-flight statuses (mutually exclusive with status).",
            "in": "query",
            "name": "in_flight",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "When true, filter to non-terminal in-flight statuses (mutually exclusive with status).",
              "title": "In Flight"
            }
          },
          {
            "description": "When true, filter to FAILED sagas only (mutually exclusive with in_flight and status).",
            "in": "query",
            "name": "failed",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "When true, filter to FAILED sagas only (mutually exclusive with in_flight and status).",
              "title": "Failed"
            }
          },
          {
            "description": "Include total matching row count.",
            "in": "query",
            "name": "include_total",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include total matching row count.",
              "title": "Include Total",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SagaInstanceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sagas",
        "tags": [
          "sagas"
        ]
      }
    },
...
    "/v1/sagas/{trace_id}/steps/{step_span_id}/retry-compensation": {
      "post": {
        "description": "Retry a failed or stuck compensation step.",
        "operationId": "operator_retry_compensation_v1_sagas__trace_id__steps__step_span_id__retry_compensation_post",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "step_span_id",
            "required": true,
            "schema": {
              "title": "Step Span Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Namespace",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/OperatorRecoveryRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Step or saga not found."
          },
          "409": {
            "description": "Recovery conflict (wrong status, stale claim, etc.)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Operator Retry Compensation",
        "tags": [
          "recovery"
        ]
      }
    },
    "/v1/sagas/{trace_id}/steps/{step_span_id}/retry-step": {
      "post": {
        "description": "Retry a stuck forward step (IN_PROGRESS) after automatic recovery is exhausted.",
        "operationId": "operator_retry_step_v1_sagas__trace_id__steps__step_span_id__retry_step_post",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "step_span_id",
            "required": true,
            "schema": {
              "title": "Step Span Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Namespace",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/OperatorRecoveryRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Step or saga not found."
          },
          "409": {
            "description": "Recovery conflict (wrong status, stale claim, etc.)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Operator Retry Step",
        "tags": [
          "recovery"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Local engine",
      "url": "http://127.0.0.1:8000"
    }
  ]
}